tests/visual-regression/requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.40.0
Pillow==10.1.0
pytest-html==4.1.1